# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/registrations": {
      "get": {
        "operationId": "Registrations_list",
        "summary": "Registration inbox (A02) / the customer's own (CU02)",
        "description": "`q` matches serial, customer, model code and dealer. Newest first by default.",
        "parameters": [
          {
            "name": "flag",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "EXCEPTION",
                "DUPLICATE",
                "MODEL_MISMATCH"
              ],
              "type": "string"
            }
          },
          {
            "name": "channel",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "DEALER",
                "PORTAL",
                "EMAIL",
                "ERP",
                "BULK"
              ],
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "PENDING",
                "APPROVED",
                "REJECTED"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated<RegistrationView>"
          }
        },
        "tags": [
          "registrations"
        ]
      },
      "post": {
        "operationId": "Registrations_create",
        "summary": "Register a unit (CU01, DL03)",
        "description": "Customer: always PENDING (needs purchaseDate and an invoice). Dealer, distributor, admin: approved at once unless the serial is already registered; other problems answer 422 with `rowErrors.<code>` field errors.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "serial": {
                    "type": "string"
                  },
                  "modelCode": {
                    "type": "string"
                  },
                  "installDate": {
                    "type": "string",
                    "format": "date"
                  },
                  "purchaseDate": {
                    "type": "string",
                    "format": "date"
                  },
                  "customerName": {
                    "type": "string"
                  },
                  "customerPhone": {
                    "type": "string"
                  },
                  "customerEmail": {
                    "type": "string"
                  },
                  "city": {
                    "type": "string"
                  },
                  "invoiceNumber": {
                    "type": "string"
                  },
                  "location": {
                    "type": "string"
                  },
                  "dealerId": {
                    "type": "string"
                  },
                  "attachmentIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "RegistrationView"
          }
        },
        "tags": [
          "registrations"
        ]
      }
    },
    "/api/registrations/bulk-approve": {
      "post": {
        "operationId": "Registrations_bulkApprove",
        "summary": "Approve several pending registrations",
        "description": "Skips duplicates, decided rows and unknown models.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "ids": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "`{ approved, skipped }`"
          }
        },
        "tags": [
          "registrations"
        ]
      }
    },
    "/api/registrations/{id}": {
      "get": {
        "operationId": "Registrations_get",
        "summary": "Registration (A03)",
        "description": "Includes `duplicateOf` (admins) when the serial is registered.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "RegistrationView"
          }
        },
        "tags": [
          "registrations"
        ]
      }
    },
    "/api/registrations/{id}/approve": {
      "post": {
        "operationId": "Registrations_approve",
        "summary": "Approve",
        "description": "`409 duplicate_serial`, `409 not_pending`, `409 unknown_model`.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "RegistrationView"
          }
        },
        "tags": [
          "registrations"
        ]
      }
    },
    "/api/registrations/{id}/reject": {
      "post": {
        "operationId": "Registrations_reject",
        "summary": "Reject with a reason",
        "description": "`422` validation.reasonRequired.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "reason"
                ],
                "properties": {
                  "reason": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "RegistrationView"
          }
        },
        "tags": [
          "registrations"
        ]
      }
    },
    "/api/registrations/{id}/merge": {
      "post": {
        "operationId": "Registrations_merge",
        "summary": "Merge a duplicate into the existing unit",
        "description": "`409 nothing_to_merge`.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "RegistrationView"
          }
        },
        "tags": [
          "registrations"
        ]
      }
    },
...
    "/api/bulk-imports/{id}": {
      "get": {
        "operationId": "BulkImports_get",
        "summary": "One upload with its rows",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "BulkImportView"
          }
        },
        "tags": [
          "bulk-imports"
        ]
      }
    },
    "/api/bulk-imports/{id}/rows": {
      "put": {
        "operationId": "BulkImports_resubmit",
        "summary": "Resubmit fixed rows",
        "description": "Only rows in ERROR are re-checked.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "rows": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "rowNumber": {
                          "type": "integer"
                        },
                        "values": {
                          "type": "object"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "BulkImportView"
          }
        },
        "tags": [
          "bulk-imports"
        ]
      }
    },
...
  },
  "info": {
    "title": "Warranty Management API",
    "description": "Units with part-wise warranties, registrations from every channel, complaints, service hand-off, manufacturer claims and the integration log. Contract: frontend/docs/api-contract.md.",
    "version": "0.2.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "wms_refresh"
      }
    },
...
  }
}